Refactor js code
